requirements.txt
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
idna==2.8
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.5
